omit group in variant and member serialization
