+++ extensions/scarb-doc/src/types.rs
@@ -1,4 +1,5 @@
 pub mod crate_type;
 pub mod groups;
+pub mod item_data;
 pub mod module_type;
 pub mod other_types;
